+++ src/containerapp/azext_containerapp/commands.py
@@ -130,8 +130,11 @@ def load_command_table(self, _):
         g.custom_show_command('show', 'show_dapr_component_resiliency')
         g.custom_show_command('list', 'list_dapr_component_resiliencies')
 
-    with self.command_group('containerapp env telemetry', is_preview=True) as g:
-        g.custom_command('set', 'set_environment_telemetry', supports_no_wait=True, exception_handler=ex_handler_factory())
+    with self.command_group('containerapp env telemetry data-dog', is_preview=True) as g:
+        g.custom_command('set', 'set_environment_telemetry_data_dog', supports_no_wait=True, exception_handler=ex_handler_factory())
+
+    with self.command_group('containerapp env telemetry app-insights', is_preview=True) as g:
+        g.custom_command('set', 'set_environment_telemetry_app_insights', supports_no_wait=True, exception_handler=ex_handler_factory())
 
     with self.command_group('containerapp github-action') as g:
         g.custom_command('add', 'create_or_update_github_action', exception_handler=ex_handler_factory())