+++ src/containerapp/azext_containerapp/containerapp_env_telemetry_decorator.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-from azure.cli.command_modules.containerapp._utils import safe_set
+from azure.cli.command_modules.containerapp._utils import safe_set, safe_get
 from knack.log import get_logger
 from knack.util import CLIError
 from ._client_factory import handle_raw_exception
@@ -10,36 +10,28 @@ from azure.cli.command_modules.containerapp.base_resource import BaseResource
 from azure.cli.core.commands import AzCliCommand
 from typing import Any, Dict
 
-
+DATA_DOG_DEST = 'dataDog'
+APP_INSIGHTS_DEST = 'appInsights'
 logger = get_logger(__name__)
 
 
-class ContainerappEnvTelemetryPreviewSetDecorator(BaseResource):
+class ContainerappEnvTelemetryDataDogPreviewSetDecorator(BaseResource):
 
     def __init__(self, cmd: AzCliCommand, client: Any, raw_parameters: Dict, models: str):
         super().__init__(cmd, client, raw_parameters, models)
         self.managed_env_def = {}
-
-    def get_argument_app_insights_connection_string(self):
-        return self.get_param("app_insights_connection_string")
-    
-    def get_argument_open_telemetry_include_system_telemetry(self):
-        return self.get_param("open_telemetry_include_system_telemetry")
     
     def get_argument_open_telemetry_data_dog_site(self):
-        return self.get_param("open_telemetry_data_dog_site")
+        return self.get_param("site")
     
     def get_argument_open_telemetry_data_dog_key(self):
-        return self.get_param("open_telemetry_data_dog_key")
+        return self.get_param("key")
     
-    def get_argument_open_telemetry_traces_destinations(self):
-        return self.get_param("open_telemetry_traces_destinations")
+    def get_argument_enable_open_telemetry_traces(self):
+        return self.get_param("enable_open_telemetry_traces")
     
-    def get_argument_open_telemetry_logs_destinations(self):
-        return self.get_param("open_telemetry_logs_destinations")
-
-    def get_argument_open_telemetry_metrics_destinations(self):
-        return self.get_param("open_telemetry_metrics_destinations")
+    def get_argument_enable_open_telemetry_metrics(self):
+        return self.get_param("enable_open_telemetry_metrics")
 
     def construct_payload(self):
         try:
@@ -50,17 +42,18 @@ class ContainerappEnvTelemetryPreviewSetDecorator(BaseResource):
         # General setup
         safe_set(self.managed_env_def, "location", value=r["location"])  # required for API
 
-        self.set_up_app_insights_connection_string()
+        existing_managed_env_def = None
+        # Get containerapp env properties of CA we are updating
+        try:
+            existing_managed_env_def = self.client.show(cmd=self.cmd, resource_group_name=self.get_argument_resource_group_name(), name=self.get_argument_name())
+        except:
+            pass
+
         self.set_up_open_telemetry_data_dog_site()
         self.set_up_open_telemetry_data_dog_key()
-        self.set_up_open_telemetry_traces_destinations()
-        self.set_up_open_telemetry_logs_destinations()
-        self.set_up_open_telemetry_metrics_destinations()
+        self.set_up_open_telemetry_traces_destinations(existing_managed_env_def)
+        self.set_up_open_telemetry_metrics_destinations(existing_managed_env_def)
 
-    def set_up_app_insights_connection_string(self):
-        if self.get_argument_app_insights_connection_string() and self.get_argument_app_insights_connection_string() is not None:
-            safe_set(self.managed_env_def, "properties", "appInsightsConfiguration", "connectionString", value=self.get_argument_app_insights_connection_string())
-    
     def set_up_open_telemetry_data_dog_site(self):
         if self.get_argument_open_telemetry_data_dog_site() and self.get_argument_open_telemetry_data_dog_site() is not None:
             safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "destinationsConfiguration", "dataDogConfiguration", "site", value=self.get_argument_open_telemetry_data_dog_site())
@@ -69,17 +62,107 @@ class ContainerappEnvTelemetryPreviewSetDecorator(BaseResource):
         if self.get_argument_open_telemetry_data_dog_key() and self.get_argument_open_telemetry_data_dog_key() is not None:
             safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "destinationsConfiguration", "dataDogConfiguration", "key", value=self.get_argument_open_telemetry_data_dog_key())
 
-    def set_up_open_telemetry_traces_destinations(self):
-        if self.get_argument_open_telemetry_traces_destinations() and self.get_argument_open_telemetry_traces_destinations() is not None:
-            safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=self.get_argument_open_telemetry_traces_destinations())
+    def set_up_open_telemetry_traces_destinations(self, existing_managed_env_def):
+        existing_traces = safe_get(existing_managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations")
+        if existing_traces is None:
+            if self.get_argument_enable_open_telemetry_traces():
+                existing_traces = [DATA_DOG_DEST]
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=existing_traces)
+        else:
+            if self.get_argument_enable_open_telemetry_traces() and DATA_DOG_DEST not in existing_traces:
+                existing_traces.append(DATA_DOG_DEST) 
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=existing_traces)
+            elif not self.get_argument_enable_open_telemetry_traces() and DATA_DOG_DEST in existing_traces:
+                existing_traces.remove(DATA_DOG_DEST)
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=existing_traces)
 
-    def set_up_open_telemetry_logs_destinations(self):
-        if self.get_argument_open_telemetry_logs_destinations() and self.get_argument_open_telemetry_logs_destinations() is not None:
-            safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "logsConfiguration", "destinations", value=self.get_argument_open_telemetry_logs_destinations())
+    def set_up_open_telemetry_metrics_destinations(self, existing_managed_env_def):
+        existing_metrics = safe_get(existing_managed_env_def, "properties", "openTelemetryConfiguration", "metricsConfiguration", "destinations")
+        if existing_metrics is None:
+            if self.get_argument_enable_open_telemetry_metrics():
+                existing_metrics = [DATA_DOG_DEST]
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "metricsConfiguration", "destinations", value=existing_metrics)
+        else:
+            if self.get_argument_enable_open_telemetry_traces() and DATA_DOG_DEST not in existing_metrics:
+                existing_metrics.append(DATA_DOG_DEST)
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "metricsConfiguration", "destinations", value=existing_metrics)
+            elif not self.get_argument_enable_open_telemetry_traces() and DATA_DOG_DEST in existing_metrics:
+                existing_metrics.remove(DATA_DOG_DEST)
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "metricsConfiguration", "destinations", value=existing_metrics)
 
-    def set_up_open_telemetry_metrics_destinations(self):
-        if self.get_argument_open_telemetry_metrics_destinations() and self.get_argument_open_telemetry_metrics_destinations() is not None:
-            safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "metricsConfiguration", "destinations", value=self.get_argument_open_telemetry_metrics_destinations())
+    def update(self):
+        try:
+            return self.client.update(cmd=self.cmd, resource_group_name=self.get_argument_resource_group_name(),
+                                      name=self.get_argument_name(), managed_environment_envelope=self.managed_env_def, no_wait=self.get_argument_no_wait())
+        except Exception as e:
+            handle_raw_exception(e)
+
+class ContainerappEnvTelemetryAppInsightsPreviewSetDecorator(BaseResource):
+
+    def __init__(self, cmd: AzCliCommand, client: Any, raw_parameters: Dict, models: str):
+        super().__init__(cmd, client, raw_parameters, models)
+        self.managed_env_def = {}
+    
+    def get_argument_open_telemetry_app_insights_connection_string(self):
+        return self.get_param("connection_string")
+    
+    def get_argument_enable_open_telemetry_traces(self):
+        return self.get_param("enable_open_telemetry_traces")
+    
+    def get_argument_enable_open_telemetry_logs(self):
+        return self.get_param("enable_open_telemetry_logs")
+
+    def construct_payload(self):
+        try:
+            r = self.client.show(cmd=self.cmd, resource_group_name=self.get_argument_resource_group_name(), name=self.get_argument_name())
+        except CLIError as e:
+            handle_raw_exception(e)
+
+        # General setup
+        safe_set(self.managed_env_def, "location", value=r["location"])  # required for API
+
+        existing_managed_env_def = None
+        # Get containerapp env properties of CA we are updating
+        try:
+            existing_managed_env_def = self.client.show(cmd=self.cmd, resource_group_name=self.get_argument_resource_group_name(), name=self.get_argument_name())
+        except:
+            pass
+
+        self.set_up_open_telemetry_open_telemetry_app_insights_connection_string()
+        self.set_up_open_telemetry_traces_destinations(existing_managed_env_def)
+        self.set_up_open_telemetry_logs_destinations(existing_managed_env_def)
+
+    def set_up_open_telemetry_open_telemetry_app_insights_connection_string(self):
+        if self.get_argument_open_telemetry_app_insights_connection_string() and self.get_argument_open_telemetry_app_insights_connection_string() is not None:
+            safe_set(self.managed_env_def, "properties", "appInsightsConfiguration", "connectionString", value=self.get_argument_open_telemetry_app_insights_connection_string())
+
+    def set_up_open_telemetry_traces_destinations(self, existing_managed_env_def):
+        existing_traces = safe_get(existing_managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations")
+        if existing_traces is None:
+            if self.get_argument_enable_open_telemetry_traces():
+                existing_traces = [APP_INSIGHTS_DEST]
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=existing_traces)
+        else:
+            if self.get_argument_enable_open_telemetry_traces() and APP_INSIGHTS_DEST not in existing_traces:
+                existing_traces.append(APP_INSIGHTS_DEST) 
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=existing_traces)
+            elif not self.get_argument_enable_open_telemetry_traces() and APP_INSIGHTS_DEST in existing_traces:
+                existing_traces.remove(APP_INSIGHTS_DEST)
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "tracesConfiguration", "destinations", value=existing_traces)
+
+    def set_up_open_telemetry_logs_destinations(self, existing_managed_env_def):
+        existing_logs = safe_get(existing_managed_env_def, "properties", "openTelemetryConfiguration", "logsConfiguration", "destinations")
+        if existing_logs is None:
+            if self.get_argument_enable_open_telemetry_logs():
+                existing_logs = [APP_INSIGHTS_DEST]
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "logsConfiguration", "destinations", value=existing_logs)
+        else:
+            if self.get_argument_enable_open_telemetry_logs() and APP_INSIGHTS_DEST not in existing_logs:
+                existing_logs.append(APP_INSIGHTS_DEST) 
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "logsConfiguration", "destinations", value=existing_logs)
+            elif not self.get_argument_enable_open_telemetry_logs() and APP_INSIGHTS_DEST in existing_logs:
+                existing_logs.remove(APP_INSIGHTS_DEST)
+                safe_set(self.managed_env_def, "properties", "openTelemetryConfiguration", "logsConfiguration", "destinations", value=existing_logs)
 
     def update(self):
         try: